+++ src/interfaces/DetailProps.ts
@@ -1,0 +1,3 @@
+export interface DetailProps {
+    title:string , description:string , rank:number
+  }
