+++ src/pages/detail-Item/components/detail/Detail.tsx
@@ -1,16 +1,14 @@
 import { Box } from "@mui/material"
-import Title from "../Title";
 import Size from "./components/size/Size";
 import DescriptionProduct from "./components/descriptionProduct/DescriptionProduct";
 import Hr from "./components/Hr";
 import ProductDetail from "./components/ProductDetail/ProductDetail";
+import Title from "../../../../components/Title";
+import { DetailProps } from "../../../../interfaces/DetailProps";
 
 
 
-
-export default function Detail({title , description , rank} : {title:string , description:string , rank:number}) {
-
-
+export default function Detail({title , description , rank} : DetailProps) {
   return (
     <Box>
         <Title fontSize={1.3} title={title} />
