+++ src/pages/detail-Item/components/detail/components/descriptionProduct/DescriptionProduct.tsx
@@ -1,4 +1,4 @@
-import Title from '../../../Title'
+import Title from '../../../../../../components/Title'
 import Description from './components/Description'
 
 export default function DescriptionProduct() {
