+++ src/pages/intro/components/Title.tsx
@@ -17,3 +17,4 @@ export default function Title() {
       </Box> 
   )
 }
+
